Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -2,8 +2,11 @@ package com.educandoweb.course.services;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exeptions.DatabaseException;
 import com.educandoweb.course.services.exeptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +32,17 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+       try {
+           if (!repository.existsById(id)) {
+               System.out.println("Error: User not found!");
+               throw new ResourceNotFoundException(id);
+           }
+           repository.deleteById(id);
+       } catch (EmptyResultDataAccessException e) {
+           throw new ResourceNotFoundException(id);
+       } catch (DataIntegrityViolationException e) {
+           throw new DatabaseException(e.getMessage());
+       }
     }
 
     public User update(Long id, User obj) {
